feat(resume): launch always starts new; drop the manual-launch resume dialog (phase 1)
The finalized UX drops the launch-time resume prompt: launching a saved config
ALWAYS starts a NEW session immediately, with a fresh id, even when the detached
registry holds left-running remotes for it. Resume moves to its own surface (the
"Remote Resumable" section, a later phase -- NOT built here).

DELETED
- src/renderer/components/ResumeSessionDialog.tsx -- the Resume / Start fresh /
  Cancel modal, and its <ResumeSessionDialog /> mount in App.tsx.
- src/renderer/stores/resumeLaunchStore.ts -- the one-slot "parked launch" store
  that existed only to feed that dialog.
- The gate in useLaunchConfig: no registry lookup, no resumableRemotesForConfig
  call, no launch-time refreshDetachedLiveness, no parked launch. The hook is
  back to blocked-check + launch.
- tests/unit/renderer/resume-launch-gate.test.tsx, replaced by
  tests/unit/renderer/launch-always-new.test.tsx.

KEPT (unchanged and still wired)
- The detachedRemotes registry + its session-state persistence round-trip, so a
  left-running remote survives an app restart.
- Leave running recording an entry; End remote dropping it (sshCloseStore).
- The liveness layer end to end: ssh-liveness.ts, probeTmuxLive, the
  SSH_CHECK_DETACHED_LIVE IPC, livenessStore (refreshDetachedLiveness /
  refreshAllDetachedLiveness / probeGoneSessions) and the pure reconcile helpers.
- The amber SshReattachBadge counter on ConfigRow + QuickStartPanel, refreshed
  from the Sidebar effect (the launch-time probe was not its only trigger).
- SshReattachGoneNotice and the app-restart probe path -- auto-reattach on app
  start is untouched, still silent, still no registry involvement.
- Reattach-by-original-id: buildLaunchSession/useLaunchSessionAction still accept
  sessionId + reconnect, for the resume surface to drive.
- resumableRemotesForConfig / offerableEntries / hasUnverifiedOffer /
  describeDetachedAge, with comments retargeted from "the dialog" to the resume
  surface that will consume them.

New tests pin launch-always-new: a config with one or several matching detached
entries spawns exactly one NEW session whose id is never a registry id, nothing
is parked, and the registry is left untouched. Verified discriminating by
temporarily reintroducing the gate (4 of them fail). Reattach plumbing pinned
separately.

typecheck clean; full vitest 9846 passed / 1 failed, the failure being an
unrelated load-flake in tests/unit/pty-buffered-write-ordering.test.ts that
passes in isolation.
